requirements.txt
django
psycopg2==2.7.5
celery==4.2.0
flower==0.9.2
djangorestframework